homework_02 completed (OOP base)

+++ homework_02/car.py
@@ -5,7 +5,7 @@ from homework_02 import engine
 class Car(base.Vehicle):
 
     def __init__(self,
-                 weight: float = 1300,
+                 weight: int = 1300,
                  fuel: float = 46,
                  fuel_consumption: float = 8.2,
                  car_engine: engine.Engine = engine.Engine(1.6, 4),
